# Conflicts:
#	Neo/Shadow In The Shade/Assets/02.Sprites/Effect/damagedEffect_critical_sheet.png.meta
#	Neo/Shadow In The Shade/Assets/02.Sprites/Effect/damagedEffect_sheet.png.meta

+++ Neo/Shadow In The Shade/Assets/01.Scripts/Enemy/Boss/Boss_Dice.cs
@@ -80,7 +80,7 @@ public class Boss_Dice : Enemy
         base.CheckHP();
     }
 
-    public override void GetHit(int damage)
+    public override void GetHit(float damage)
     {
         base.GetHit(damage);
     }
